kirby api reworked to composable, only named exports used, use* naming for composables
