--- vue-json/src/composables/root.js
@@ -1,3 +1,0 @@
-import { getCurrentInstance } from 'vue'
-
-export default () => getCurrentInstance().ctx
