+++ vue-json/src/composables/use-root.js
@@ -1,0 +1,3 @@
+import { getCurrentInstance } from 'vue'
+
+export const useRoot = () => getCurrentInstance().ctx
