+++ vue-json/src/main.js
@@ -1,20 +1,16 @@
 import { createApp } from 'vue'
+import { useRouter } from './router'
+import { useKirby } from './composables/use-kirby'
 import App from './App'
-import Router from './router'
-import KirbyApi from './plugins/kirby-api'
-
-const apiUrl =
-  process.env.NODE_ENV === 'production' ? process.env.VUE_APP_KIRBY_URL : window.location.origin + process.env.VUE_APP_BASE_URL.slice(0, -1)
 
 // self invoke async initialization
 ;(async () => {
-  const api = KirbyApi.init(apiUrl)
-  const site = await api.getSite()
-  const router = await Router.init(site)
+  const { getSite } = useKirby()
+  const site = await getSite()
+  const router = await useRouter(site)
   const app = createApp(App)
 
   // globals
-  app.config.globalProperties.$api = api
   app.config.globalProperties.$site = site
 
   app.use(router).mount('#app')
