+++ vue-json/src/router/index.js
@@ -1,32 +1,30 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Default from '../views/Default'
 
-const capitalize = string => string.charAt(0).toUpperCase() + string.slice(1)
+export const useRouter = async site => {
+  const capitalize = string => string.charAt(0).toUpperCase() + string.slice(1)
 
-export default {
-  init: async site => {
-    // published pages routes
-    const routes = site.children.flatMap(page => [
-      {
-        path: '/' + page.id,
-        component: () => import(`../views/${capitalize(page.template)}`).catch(() => Default)
-      },
-      ...page.children.map(child => ({
-        path: '/' + child.id,
-        component: () => import(`../views/${capitalize(child.template)}`).catch(() => Default)
-      }))
-    ])
+  // published pages routes
+  const routes = site.children.flatMap(page => [
+    {
+      path: '/' + page.id,
+      component: () => import(`../views/${capitalize(page.template)}`).catch(() => Default)
+    },
+    ...page.children.map(child => ({
+      path: '/' + child.id,
+      component: () => import(`../views/${capitalize(child.template)}`).catch(() => Default)
+    }))
+  ])
 
-    // home route / instead of /home
-    routes.find(route => route.path === '/home').path = '/'
-    routes.push({ path: '/home', redirect: '/' })
+  // home route / instead of /home
+  routes.find(route => route.path === '/home').path = '/'
+  routes.push({ path: '/home', redirect: '/' })
 
-    // catch all fallback
-    routes.push({ path: '/:pathMatch(.*)*', component: Default })
+  // catch all fallback
+  routes.push({ path: '/:pathMatch(.*)*', component: Default })
 
-    return createRouter({
-      history: createWebHistory(process.env.VUE_APP_BASE_URL),
-      routes
-    })
-  }
+  return createRouter({
+    history: createWebHistory(process.env.VUE_APP_BASE_URL),
+    routes
+  })
 }
