+++ vue-json/src/views/About.vue
@@ -40,8 +40,8 @@
 </template>
 
 <script>
+import { usePage } from '../composables/use-page'
 import Intro from '../components/Intro'
-import usePage from '../composables/page'
 
 export default {
   name: 'About',
