+++ vue-json/src/views/Default.vue
@@ -7,8 +7,8 @@
 </template>
 
 <script>
+import { usePage } from '../composables/use-page'
 import Intro from '../components/Intro'
-import usePage from '../composables/page'
 
 export default {
   name: 'Default',
