+++ vue-json/src/views/Note.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import usePage from '../composables/page'
+import { usePage } from '../composables/use-page'
 
 export default {
   name: 'Note',
